# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FULL SWAP API - Real-time AI video transformation platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered — email verification required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-email:
    post:
      operationId: verifyEmail
      tags: [auth]
      summary: Verify email with 5-digit PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                pin:
                  type: string
              required: [email, pin]
      responses:
        "200":
          description: Email verified — token returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid or expired PIN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/resend-verification:
    post:
      operationId: resendVerification
      tags: [auth]
      summary: Resend email verification PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required: [email]
      responses:
        "200":
          description: PIN resent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password using email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordBody"
      responses:
        "200":
          description: Password reset successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Email not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      deprecated: true
      summary: "[DEPRECATED] User login — returns 410 Gone"
      description: >
        User accounts have been removed. This endpoint always returns
        410 Gone with code ACCOUNTS_REMOVED. Admin access uses
        POST /admin/login. Regular users authenticate via license key.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "410":
          description: Gone — user accounts have been removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /chat:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message to support
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
              required: [message]
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
...
  /chat/admin/thread/{userId}:
    get:
      operationId: getAdminChatThread
      tags: [chat]
      summary: Get a specific user's chat thread (admin)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
  /chat/admin/reply/{userId}:
    post:
      operationId: sendAdminChatReply
      tags: [chat]
      summary: Send admin reply to a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
              required: [message]
      responses:
        "201":
          description: Reply sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    RegisterResponse:
      type: object
      properties:
        requiresVerification:
          type: boolean
        email:
          type: string
        message:
          type: string
      required: [requiresVerification, email]
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        sender:
          type: string
          enum: [user, admin]
        message:
          type: string
        readByAdmin:
          type: integer
        readByUser:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, userId, sender, message, readByAdmin, readByUser, createdAt]
...
    RegisterBody:
      type: object
      properties:
        email:
          type: string
          format: email
        username:
          type: string
          minLength: 3
          maxLength: 20
          pattern: '^[a-zA-Z0-9_]+$'
        password:
          type: string
          minLength: 8
        browserFingerprint:
          type: string
          description: Client-side browser fingerprint hash for abuse prevention
      required: [email, username, password]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    ResetPasswordBody:
      type: object
      properties:
        email:
          type: string
          format: email
        newPassword:
          type: string
          minLength: 8
      required: [email, newPassword]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: string
        membership:
          type: string
          enum: [active, suspended, free_trial]
        freeSecondsRemaining:
          type: integer
        totalMinutesPurchased:
          type: integer
        totalMinutesUsed:
          type: number
        createdAt:
          type: string
          format: date-time
        avatarUrl:
          type: string
          nullable: true
          description: Optional profile avatar URL
        isSubAdmin:
          type: integer
          enum: [0, 1]
          description: "1 if user has sub-admin privileges, 0 otherwise"
      required: [id, email, username, membership, freeSecondsRemaining, totalMinutesPurchased, totalMinutesUsed, createdAt]
...